Sensors: (name, thing) is unique
the tuple (name, thing) is now used to identify a sensor, in contrast to name only.
when the name of the thing is not found, it is created with empty owner array, the given name and the sensor in its sensors array.
when the thing already exists, the program tries to add the sensor to the things sensors array.
if the sensor is already in the array, nothing happens.

deleting a sensor requires to give the thing and the name of the sensor.
the sensor is removed from the thing but the thing still exists afterwards.